继续完成剩下的分析接口的编写，增加日志功能，前端界面的补全

+++ backend/app/schemas/customer.py
@@ -1,10 +1,9 @@
-from decimal import Decimal
 from pydantic import BaseModel
 
 
 class CustomerRankItem(BaseModel):
     customer_unique_id: str
-    sales: Decimal
+    sales: float
     count_order: int
